tonai-server/pyproject.toml
[tool.poetry]
name = "TonAIStarkAgent"
version = "0.1.0"
description = ""
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.3.22"
langchain = "^0.3.14"
langchain-openai = "^0.3.11"
python-dotenv = "^1.0.1"
fastapi = "^0.115.5"
uvicorn = "^0.34.0"
sse-starlette = ">=2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = "app.server:main"
